# docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /v2/auth/register:
    post:
      operationId: AuthController_register
      summary: "Register a new user (multipart: optional profilePicture)"
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        201:
          description: ""
      tags:
        - Auth
  /v2/auth/login:
    post:
      operationId: AuthController_login
      summary: Login and receive JWT token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        200:
          description: ""
      tags:
        - Auth
  /v2/auth/logout:
    post:
      operationId: AuthController_logout
      summary: Logout (client should discard JWT)
      parameters: []
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
  /v2/auth/change-password:
    post:
      operationId: AuthController_changePassword
      summary: Change password (requires current password verification)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
  /v2/auth/me:
    get:
      operationId: AuthController_getMe
      summary: Get current authenticated user info
      parameters: []
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
  /v2/auth/roles:
    get:
      operationId: AuthController_getRoles
      summary: Get list of available roles
      parameters: []
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
  /v2/auth/users:
    get:
      operationId: AuthController_listUsers
      summary: Danh sách tài khoản để duyệt (admin). status=pending là hàng đợi guest chờ gắn member.
      parameters:
        - 
          name: status
          required: false
          in: query
          schema:
            enum:
              - pending
              - linked
              - all
            type: string
        - 
          name: role
          required: false
          in: query
          schema:
            enum:
              - guest
              - member
              - editor
              - admin
            type: string
        - 
          name: page
          required: false
          in: query
          schema:
            type: number
        - 
          name: pageSize
          required: false
          in: query
          schema:
            type: number
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
  /v2/auth/users/{userId}/roles:
    put:
      operationId: AuthController_assignRoles
      summary: Đổi role của một tài khoản (admin). Không tự đổi role của chính mình.
      parameters:
        - 
          name: userId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRolesDto"
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
  /v2/auth/users/{userId}/link-member:
    post:
      operationId: AuthController_linkMember
      summary: Gắn tài khoản vào một member CÓ SẴN (admin). Guest được nâng lên member.
      parameters:
        - 
          name: userId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkMemberDto"
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
    delete:
      operationId: AuthController_unlinkMember
      summary: Gỡ liên kết tài khoản ↔ member (admin). Member bị hạ về guest.
      parameters:
        - 
          name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
      tags:
        - Auth
      security:
        - 
          bearer: []
...
  /v2/events/{id}:
    get:
      operationId: EventsController_getById
      summary: Get event by ID
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
    put:
      operationId: EventsController_update
      summary: Update event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
    delete:
      operationId: EventsController_delete
      summary: Delete event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
...
  /v2/anniversaries:
    get:
      operationId: AnniversariesController_getAnniversaries
      summary: Get anniversaries (filter by member_id, month)
      parameters:
        - 
          name: member_id
          required: false
          in: query
          schema:
            type: string
        - 
          name: month
          required: false
          in: query
          schema:
            type: number
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnniversaryResponseDto"
      tags:
        - Anniversaries (Ngày giỗ)
      security:
        - 
          bearer: []
    post:
      operationId: AnniversariesController_create
      summary: Create anniversary (optional member link)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnniversaryDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnniversaryResponseDto"
      tags:
        - Anniversaries (Ngày giỗ)
      security:
        - 
          bearer: []
  /v2/anniversaries/upcoming:
    get:
      operationId: AnniversariesController_getUpcoming
      summary: Get upcoming anniversaries (next 30 days)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnniversaryResponseDto"
      tags:
        - Anniversaries (Ngày giỗ)
      security:
        - 
          bearer: []
  /v2/anniversaries/{id}:
    get:
      operationId: AnniversariesController_getById
      summary: Get anniversary by ID
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnniversaryResponseDto"
      tags:
        - Anniversaries (Ngày giỗ)
      security:
        - 
          bearer: []
    put:
      operationId: AnniversariesController_update
      summary: Update anniversary
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnniversaryDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnniversaryResponseDto"
      tags:
        - Anniversaries (Ngày giỗ)
      security:
        - 
          bearer: []
    delete:
      operationId: AnniversariesController_delete
      summary: Delete anniversary
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Anniversaries (Ngày giỗ)
      security:
        - 
          bearer: []
...
  /v2/articles:
    get:
      operationId: ArticlesController_getArticles
      summary: "List articles (filter: category, year, search) — public"
      parameters:
        - 
          name: category
          required: false
          in: query
          schema:
            type: string
        - 
          name: year
          required: false
          in: query
          schema:
            type: number
        - 
          name: search
          required: false
          in: query
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleResponseDto"
      tags:
        - Articles (Tin tức dòng họ)
      security:
        - 
          bearer: []
    post:
      operationId: ArticlesController_create
      summary: Create article
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponseDto"
      tags:
        - Articles (Tin tức dòng họ)
      security:
        - 
          bearer: []
  /v2/articles/{id}:
    get:
      operationId: ArticlesController_getById
      summary: Get article by ID (public)
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponseDto"
      tags:
        - Articles (Tin tức dòng họ)
      security:
        - 
          bearer: []
    put:
      operationId: ArticlesController_update
      summary: Update article
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponseDto"
      tags:
        - Articles (Tin tức dòng họ)
      security:
        - 
          bearer: []
    delete:
      operationId: ArticlesController_delete
      summary: Delete article
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Articles (Tin tức dòng họ)
      security:
        - 
          bearer: []
  /v2/articles/{id}/view:
    post:
      operationId: ArticlesController_incrementView
      summary: Increment article view count (public)
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: View counted
      tags:
        - Articles (Tin tức dòng họ)
      security:
        - 
          bearer: []
info:
  title: Family Tree API v2
  description: "## Vietnamese Family Tree Management API\n\n**Base URL:** `http://localhost:3002/v2`\n\n**Authentication:** Bearer JWT token (same as backend v1)\n\n### Modules\n- **Members** — CRUD thành viên + profile + avatar (async upload)\n- **Relationships** — Quan hệ mới (BIOLOGICAL/ADOPTED/SPOUSE) + tìm tổ tiên/con cháu\n- **Tree** — Cây gia phả full (Redis cache 1h) + subtree 4 thế hệ\n- **Anniversaries** — Ngày giỗ (filter by member, month, upcoming)\n- **Events** — Sự kiện dòng họ + notification queue\n- **Media** — Thư viện media: upload mọi loại (ảnh nén lossless bằng sharp; video/audio/tài liệu upload thẳng) → Vercel Blob; phân trang/lọc/tìm kiếm + thống kê + album\n- **Graves** — Mộ phần với tọa độ GPS + tìm kiếm gần nhất\n\n### Queue Jobs (QStash + Redis)\n| Queue | Trigger | Action |\n|-------|---------|--------|\n| avatar-upload | Create/Update member với file | Upload → Vercel Blob → cập nhật avatar_url |\n| media-process | Upload media | Ảnh: sharp nén lossless; khác: upload thẳng → Vercel Blob (off-request) |\n| report-generate | Create/Delete member | Tính stats → lưu Redis |\n| notification | New member/relationship/event | Log (Phase 1) |\n"
  version: 2.0.0
  contact:
    name: Family Tree Team
    url: ""
    email: ""
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
      description: Dán CHỈ token thô (lấy từ POST /v2/auth/login), KHÔNG kèm tiền tố "Bearer ".
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
          minLength: 6
        fullName:
          type: string
          example: Nguyen Van A
        gender:
          type: string
          example: male
        birthDate:
          type: string
          example: 1990-01-15
        deathDate:
          type: string
          example: 2020-05-10
        generation:
          type: number
          example: 3
        occupation:
          type: string
          example: Software Engineer
        address:
          type: string
          example: Hanoi, Vietnam
        biography:
          type: string
          example: A brief biography...
      required:
        - email
        - password
        - fullName
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
      required:
        - email
        - password
    ChangePasswordDto:
      type: object
      properties:
        currentPassword:
          type: string
          example: oldPassword123
        newPassword:
          type: string
          example: newPassword456
          minLength: 6
      required:
        - currentPassword
        - newPassword
    AssignRolesDto:
      type: object
      properties:
        roles:
          type: array
          example:
            - editor
          items:
            type: string
            enum:
              - guest
              - member
              - editor
              - admin
      required:
        - roles
    LinkMemberDto:
      type: object
      properties:
        memberId:
          type: string
          example: 3f2504e0-4f89-11d3-9a0c-0305e82c3301
      required:
        - memberId
...
    MemberTreeBriefDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: object
          nullable: true
          example: Chi nhánh Phú Thọ
      required:
        - id
...
    MemberResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        private_id:
          type: object
          nullable: true
        name:
          type: string
          example: Nguyễn Văn A
        normalized_name:
          type: object
          nullable: true
          description: Tên đã bỏ dấu để tìm kiếm
        avatar_url:
          type: object
          nullable: true
          format: uri
        gender:
          type: string
          nullable: true
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: object
          nullable: true
          example: 1990-01-01
        deathDate:
          type: object
          nullable: true
          example: 2020-12-31
        tree_id:
          type: object
          nullable: true
          format: uuid
          description: Chi nhánh — FK tới Tree
        generation:
          type: number
          nullable: true
          example: 5
          description: "Thế hệ (giá trị HIỆU LỰC, đã tính sẵn) = `profile.generation` nếu nhập tay, ngược lại suy ra từ quan hệ cha-con/vợ-chồng. Cập nhật bởi job nền `generation-recompute`. Muốn biết là nhập tay hay suy ra: `profile.generation != null`."
        generation_updated_at:
          type: string
          nullable: true
          format: date-time
          description: Lần cuối job nền ghi lại `generation`
        created_at:
          type: string
          format: date-time
        profile:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/ProfileResponseDto"
        tree:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/MemberTreeBriefDto"
      required:
        - id
        - name
        - created_at
...
    EventResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Giỗ tổ dòng họ
        description:
          type: object
          nullable: true
        images:
          description: Ảnh sự kiện
          type: array
          items:
            type: string
            format: uri
        highlight:
          type: boolean
          default: false
          description: Nổi bật trên trang chủ
        date:
          type: string
          nullable: true
          format: date-time
        end_date:
          type: string
          nullable: true
          format: date-time
          description: Ngày kết thúc
        location:
          type: object
          nullable: true
          example: Nhà thờ họ
        category:
          type: object
          nullable: true
          example: Lễ
          description: Loại sự kiện
        isLunar:
          type: boolean
          default: false
          description: Ngày theo âm lịch
        attendeeCount:
          type: number
          example: 0
          description: Số người tham dự
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - images
        - highlight
        - isLunar
        - attendeeCount
        - created_at
        - updated_at
...
    UpdateEventDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          format: date-time
          type: string
        end_date:
          format: date-time
          type: string
        location:
          type: string
        category:
          type: string
        isLunar:
          type: boolean
        highlight:
          type: boolean
        images:
          type: array
          items:
            type: string
        existingImages:
          description: Ảnh hiện có được giữ lại (URL); ảnh mới upload sẽ nối vào sau
          type: array
          items:
            type: string
...
    AnniversaryResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Giỗ cụ Nguyễn Văn A
        date:
          type: string
          format: date-time
        description:
          type: object
          nullable: true
        member_id:
          type: object
          nullable: true
          format: uuid
        cemetery_id:
          type: object
          nullable: true
          format: uuid
          description: Mộ/nghĩa trang liên kết
        isLunar:
          type: boolean
          default: false
          description: Ngày giỗ theo âm lịch
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        member:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/MemberResponseDto"
        cemetery:
          type: object
          nullable: true
          description: Mộ/nghĩa trang liên kết
      required:
        - id
        - title
        - date
        - isLunar
        - created_at
        - updated_at
    CreateAnniversaryDto:
      type: object
      properties:
        title:
          type: string
          example: Giỗ Ông Nội
        date:
          format: date-time
          type: string
          example: 2024-03-15T00:00:00.000Z
        description:
          type: string
        member_id:
          type: string
          description: Link to a member (optional)
        cemetery_id:
          type: string
          description: Link to a cemetery/grave (optional)
        isLunar:
          type: boolean
          default: false
          description: Ngày giỗ theo âm lịch
      required:
        - title
        - date
    UpdateAnniversaryDto:
      type: object
      properties:
        title:
          type: string
        date:
          format: date-time
          type: string
        description:
          type: string
        member_id:
          type: string
        cemetery_id:
          type: string
        isLunar:
          type: boolean
...
    ArticleResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        excerpt:
          type: object
          nullable: true
        content:
          type: string
        category:
          type: string
          example: announcement
        date:
          type: string
          format: date-time
        views:
          type: number
          example: 0
        featured:
          type: boolean
          default: false
        coverUrl:
          type: object
          nullable: true
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - category
        - date
        - views
        - featured
        - created_at
        - updated_at
    CreateArticleDto:
      type: object
      properties:
        title:
          type: string
          example: Lễ giỗ Tổ họ Đỗ năm 2026
        content:
          type: string
          example: Nội dung đầy đủ của bài viết...
        category:
          type: string
          example: announcement
          description: announcement | activity | story | archive | construction | other
        excerpt:
          type: string
          example: Tóm tắt ngắn hiển thị trên thẻ.
        featured:
          type: boolean
          default: false
        coverUrl:
          type: string
          format: uri
        date:
          format: date-time
          type: string
          example: 2026-02-20T00:00:00.000Z
      required:
        - title
        - content
        - category
    UpdateArticleDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        excerpt:
          type: string
        featured:
          type: boolean
        coverUrl:
          type: string
        date:
          format: date-time
          type: string